<Feat>: make test code: mca-app-test
